Refactor route/waypoints to gps.ts
